+++ test/eigen/LOBPCG/lobpcg.cpp
@@ -5,7 +5,8 @@
 
 template <typename T, typename PRECOND>
 bool test_solve(monolish::matrix::COO<T> mat, const T exact_result,
-                const int check_ans, const T tol_ev, const T tol_res) {
+                const int check_ans, const T tol_ev, const T tol_res,
+                const std::string s) {
   monolish::matrix::CRS<T> A(mat);
   T lambda;
   monolish::vector<T> x(A.get_row());
@@ -16,7 +17,7 @@ bool test_solve(monolish::matrix::COO<T> mat, const T exact_result,
   solver.set_tol(tol_res);
   solver.set_lib(0);
   solver.set_miniter(0);
-  solver.set_maxiter(1000);
+  solver.set_maxiter(A.get_row());
 
   // precond setting
   PRECOND precond;
@@ -30,7 +31,8 @@ bool test_solve(monolish::matrix::COO<T> mat, const T exact_result,
   }
 
   if (check_ans == 1) {
-    if (ans_check<T>("LOBPCG", lambda, exact_result, tol_ev) == false) {
+    if (ans_check<T>("LOBPCG(" + s + ")", lambda, exact_result, tol_ev) ==
+        false) {
       return false;
     }
   }
@@ -42,41 +44,22 @@ bool test_tridiagonal_toeplitz(const int check_ans, const T tol_ev,
                                const T tol_res) {
   int DIM = 100;
   monolish::matrix::COO<T> COO =
-      monolish::util::tridiagonal_toeplitz_matrix<T>(DIM, 11.0, -1.0);
-  // for (std::size_t i = 0; i < COO.get_row(); ++i) {
-  //   for (std::size_t j = 0; j < COO.get_col(); ++j) {
-  //     if (i == j) { COO.insert(i, j, 2.0); }
-  //     if (i - j == 1 || j - i == 1) { COO.insert(i, j, -1.0); }
-  //     if (i == 0 && j == DIM - 1) { COO.insert(i, j, -1.0); }
-  //     if (j == 0 && i == DIM - 1) { COO.insert(i, j, -1.0); }
-  //   }
-  // }
-
-  // Calculate exact result by solving full eigenproblem
-  // monolish::matrix::Dense<T> AD(COO);
-  // monolish::vector<T> ld(AD.get_row());
-  // const char jobz = 'V';
-  // const char uplo = 'U';
-  // bool bl = monolish::lapack::syev(&jobz, &uplo, AD, ld);
-  // if (!bl) {
-  //   throw std::runtime_error("LAPACK syev failed");
-  // }
-  // T exact_result = ld[0];
-
-  // Calculate exact eigenvalue from analytic solution
+      monolish::util::tridiagonal_toeplitz_matrix<T>(DIM, 10.0, -1.0);
   T exact_result = monolish::util::tridiagonal_toeplitz_matrix_eigenvalue<T>(
-      DIM, 0, 11.0, -1.0);
+      DIM, 0, 10.0, -1.0);
 
-  return test_solve<T, PRECOND>(COO, exact_result, check_ans, tol_ev, tol_res);
+  return test_solve<T, PRECOND>(COO, exact_result, check_ans, tol_ev,
+                                tol_res * DIM, "Tridiagonal Toeplitz");
 }
 
 template <typename T, typename PRECOND>
-bool test_laplacian_1d(const int check_ans, const T tol_ev, const T tol_res) {
-  int DIM = 10000;
-  monolish::matrix::COO<T> COO = monolish::util::laplacian_matrix_1D<T>(DIM);
-  T exact_result = monolish::util::laplacian_matrix_1D_eigenvalue<T>(DIM, 0);
+bool test_frank(const int check_ans, const T tol_ev, const T tol_res) {
+  int DIM = 100;
+  monolish::matrix::COO<T> COO = monolish::util::frank_matrix<T>(DIM);
+  T exact_result = monolish::util::frank_matrix_eigenvalue<T>(DIM, 0);
 
-  return test_solve<T, PRECOND>(COO, exact_result, check_ans, tol_ev, tol_res);
+  return test_solve<T, PRECOND>(COO, exact_result, check_ans, tol_ev,
+                                tol_res * DIM, "Frank");
 }
 
 int main(int argc, char **argv) {
@@ -95,45 +78,45 @@ int main(int argc, char **argv) {
   if (test_tridiagonal_toeplitz<
           double,
           monolish::equation::none<monolish::matrix::CRS<double>, double>>(
-          check_ans, 1.0e-8, 1.0e-4) == false) {
+          check_ans, 3.0e-4, 1.0e-5) == false) {
     return 1;
   }
   if (test_tridiagonal_toeplitz<
           float, monolish::equation::none<monolish::matrix::CRS<float>, float>>(
-          check_ans, 2.0e-2, 1.0e-2) == false) {
+          check_ans, 3.0e-4, 1.0e-5) == false) {
     return 1;
   }
   if (test_tridiagonal_toeplitz<
           double,
           monolish::equation::Jacobi<monolish::matrix::CRS<double>, double>>(
-          check_ans, 1.0e-8, 1.0e-4) == false) {
+          check_ans, 3.0e-4, 1.0e-5) == false) {
     return 1;
   }
   if (test_tridiagonal_toeplitz<
           float,
           monolish::equation::Jacobi<monolish::matrix::CRS<float>, float>>(
-          check_ans, 2.0e-2, 1.0e-2) == false) {
+          check_ans, 3.0e-4, 1.0e-5) == false) {
     return 1;
   }
 
-  if (test_laplacian_1d<double, monolish::equation::none<
-                                    monolish::matrix::CRS<double>, double>>(
-          check_ans, 4.0e-3, 5.0e-3) == false) {
+  if (test_frank<double, monolish::equation::none<monolish::matrix::CRS<double>,
+                                                  double>>(check_ans, 2.0e-1,
+                                                           1.0e-5) == false) {
     return 1;
   }
-  if (test_laplacian_1d<
-          float, monolish::equation::none<monolish::matrix::CRS<float>, float>>(
-          check_ans, 2.0e-2, 1.0e-2) == false) {
+  if (test_frank<float,
+                 monolish::equation::none<monolish::matrix::CRS<float>, float>>(
+          check_ans, 2.0e-1, 1.0e-5) == false) {
     return 1;
   }
-  if (test_laplacian_1d<double, monolish::equation::Jacobi<
-                                    monolish::matrix::CRS<double>, double>>(
-          check_ans, 4.0e-3, 5.0e-3) == false) {
+  if (test_frank<double, monolish::equation::Jacobi<
+                             monolish::matrix::CRS<double>, double>>(
+          check_ans, 2.0e-1, 1.0e-5) == false) {
     return 1;
   }
-  if (test_laplacian_1d<float, monolish::equation::Jacobi<
-                                   monolish::matrix::CRS<float>, float>>(
-          check_ans, 2.0e-2, 1.0e-2) == false) {
+  if (test_frank<float, monolish::equation::Jacobi<monolish::matrix::CRS<float>,
+                                                   float>>(check_ans, 2.0e-1,
+                                                           1.0e-5) == false) {
     return 1;
   }
   return 0;